port to MonoGame 3.6

+++ Nez.Portable/Core.cs
@@ -319,7 +319,7 @@ namespace Nez
 				TimeRuler.instance.render();
 
 			#if !FNA
-			drawCalls = graphicsDevice.Metrics.DrawCount;
+			drawCalls = (ulong) graphicsDevice.Metrics.DrawCount;
 			#endif
 			#endif
 		}
